refactor: move `NavItemProps` to separate file
