+++ src/components/ui/Nav/NavItemProps.ts
@@ -1,0 +1,11 @@
+export interface NavItemProps {
+  href?: string;
+  text: string;
+  subItems?: SubItem[];
+}
+
+export interface SubItem {
+  href: string;
+  text: string;
+  description: string;
+}
