+++ src/components/ui/Nav/index.ts
@@ -1,2 +1,3 @@
 export * from './NavItem';
 export * from './NavItemMobile';
+export * from './NavItemProps';
